Fix InvalidateCacheWorker to work fine

+++ app/src/main/java/com/ashehata/dru/worker/InvalidateCacheWorker.kt
@@ -1,4 +1,4 @@
-package com.ashehata.dru.work
+package com.ashehata.dru.worker
 
 import android.content.Context
 import android.util.Log
@@ -6,24 +6,31 @@ import androidx.hilt.work.HiltWorker
 import androidx.work.CoroutineWorker
 import androidx.work.WorkerParameters
 import com.ashehata.dru.features.movies.data.local.dao.MoviesDao
+import dagger.assisted.Assisted
+import dagger.assisted.AssistedInject
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
-import javax.inject.Inject
 
 @HiltWorker
-class InvalidateCacheWorker @Inject constructor(
+class InvalidateCacheWorker @AssistedInject constructor(
+    @Assisted appContext: Context,
+    @Assisted workerParams: WorkerParameters,
     private val moviesDao: MoviesDao,
-    appContext: Context,
-    workerParams: WorkerParameters,
 ) : CoroutineWorker(appContext, workerParams) {
+
+    companion object {
+        const val NAME = "InvalidateCacheWorker"
+    }
+
     override suspend fun doWork(): Result {
         return withContext(Dispatchers.IO) {
             try {
-                Log.i("InvalidateCacheWorker", "doWork: ")
-                moviesDao.deleteAll()
+                Log.i(NAME, "doWork: ")
+                if (moviesDao.getMovies().isNotEmpty())
+                    moviesDao.deleteAll()
                 Result.success()
             } catch (e: Exception) {
-                Log.i("InvalidateCacheWorker", "Exception: " + e.localizedMessage)
+                Log.i(NAME, "Exception: " + e.localizedMessage)
                 Result.failure()
             }
         }
